warnings be gone (2/4)
(cherry picked from commit 84333e0475bc911adc16417f4ca327c975cf6c36)
(squashed with commit b4a7a2df4c28c3f32b5d877b54831d2cc5d78f81)

+++ media/libstagefright/httplive/M3UParser.cpp
@@ -162,18 +162,18 @@ status_t M3UParser::MediaGroup::selectTrack(size_t index, bool select) {
 
     if (select) {
         if (index >= mMediaItems.size()) {
-            ALOGE("track %d does not exist", index);
+            ALOGE("track %zu does not exist", index);
             return INVALID_OPERATION;
         }
-        if (mSelectedIndex == index) {
-            ALOGE("track %d already selected", index);
+        if (mSelectedIndex == (ssize_t)index) {
+            ALOGE("track %zu already selected", index);
             return BAD_VALUE;
         }
         ALOGV("selected track %d", index);
         mSelectedIndex = index;
     } else {
-        if (mSelectedIndex != index) {
-            ALOGE("track %d is not selected", index);
+        if (mSelectedIndex != (ssize_t)index) {
+            ALOGE("track %zu is not selected", index);
             return BAD_VALUE;
         }
         ALOGV("unselected track %d", index);
